Better work width detection
Get the left/right distances of work areas or AI markers for
each implement (instead of just the width) and use those to
determine the total working width.

Also, do the unfold/fold trick before getting the width so if
work areas and markers fold with the implement are in their
working position when we measure the distances.

This commit has only the width fix, offset comes next.